requirements.txt
discord.py>=2.0.0
PyNaCl>=1.4.0
PyYAML>=6.0
aiohttp>=3.8.0
numpy>=1.20.0
torch>=1.10.0
SpeechRecognition>=3.8.1
pydub>=0.25.1